app/requirements.txt
asgiref==3.8.1
Django==5.0.3
sqlparse==0.4.4

psycopg==3.1.18
psycopg2-binary==2.9.9

target365_sdk==1.8.3

cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
httplib2==0.21.0
idna==3.4
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.6.0
pyparsing==3.0.9
pytz==2022.5
requests==2.28.1
rsa==4.9
six==1.16.0
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12